Bump lemmyApi and refactor PostsViewModel Code
