+++ app/src/main/java/com/jerboa/model/CommunityViewModel.kt
@@ -2,7 +2,6 @@ package com.jerboa.model
 
 import android.content.Context
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
@@ -10,144 +9,34 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.CreationExtras
 import arrow.core.Either
-import com.jerboa.JerboaAppState
 import com.jerboa.api.API
 import com.jerboa.api.ApiState
 import com.jerboa.api.toApiState
-import com.jerboa.findAndUpdatePost
-import com.jerboa.mergePosts
 import com.jerboa.showBlockCommunityToast
-import it.vercruysse.lemmyapi.dto.SortType
 import it.vercruysse.lemmyapi.v0x19.datatypes.BlockCommunity
 import it.vercruysse.lemmyapi.v0x19.datatypes.BlockCommunityResponse
 import it.vercruysse.lemmyapi.v0x19.datatypes.CommunityId
 import it.vercruysse.lemmyapi.v0x19.datatypes.CommunityResponse
-import it.vercruysse.lemmyapi.v0x19.datatypes.CreatePostLike
-import it.vercruysse.lemmyapi.v0x19.datatypes.DeletePost
 import it.vercruysse.lemmyapi.v0x19.datatypes.FollowCommunity
 import it.vercruysse.lemmyapi.v0x19.datatypes.GetCommunity
 import it.vercruysse.lemmyapi.v0x19.datatypes.GetCommunityResponse
 import it.vercruysse.lemmyapi.v0x19.datatypes.GetPosts
-import it.vercruysse.lemmyapi.v0x19.datatypes.GetPostsResponse
-import it.vercruysse.lemmyapi.v0x19.datatypes.MarkPostAsRead
-import it.vercruysse.lemmyapi.v0x19.datatypes.PaginationCursor
-import it.vercruysse.lemmyapi.v0x19.datatypes.PostResponse
-import it.vercruysse.lemmyapi.v0x19.datatypes.PostView
-import it.vercruysse.lemmyapi.v0x19.datatypes.SavePost
 import kotlinx.coroutines.launch
 
-class CommunityViewModel(communityArg: Either<CommunityId, String>) : ViewModel() {
+class CommunityViewModel(communityArg: Either<CommunityId, String>) : PostsViewModel() {
     var communityRes: ApiState<GetCommunityResponse> by mutableStateOf(ApiState.Empty)
         private set
 
-    private var followCommunityRes: ApiState<CommunityResponse> by
-        mutableStateOf(ApiState.Empty)
-
-    var postsRes: ApiState<GetPostsResponse> by mutableStateOf(ApiState.Empty)
-        private set
-
-    private var likePostRes: ApiState<PostResponse> by mutableStateOf(ApiState.Empty)
-    private var savePostRes: ApiState<PostResponse> by mutableStateOf(ApiState.Empty)
-    private var deletePostRes: ApiState<PostResponse> by mutableStateOf(ApiState.Empty)
-    private var blockCommunityRes: ApiState<BlockCommunityResponse> by
-        mutableStateOf(ApiState.Empty)
-    private var markPostRes: ApiState<Unit> by mutableStateOf(ApiState.Empty)
-
-    var sortType by mutableStateOf(SortType.Active)
-        private set
-
-    private var page by mutableIntStateOf(1)
-    private var pageCursor: PaginationCursor? by mutableStateOf(null)
-
+    private var followCommunityRes: ApiState<CommunityResponse> by mutableStateOf(ApiState.Empty)
+    private var blockCommunityRes: ApiState<BlockCommunityResponse> by mutableStateOf(ApiState.Empty)
     private var communityId: Int? by mutableStateOf(null)
     private var communityName: String? by mutableStateOf(null)
-
-    fun updateSortType(sortType: SortType) {
-        this.sortType = sortType
-    }
-
-    private fun resetPage() {
-        page = 1
-        pageCursor = null
-    }
-
-    private fun nextPage() {
-        page += 1
-    }
-
-    private fun prevPage() {
-        page -= 1
-    }
 
     private fun getCommunity(form: GetCommunity) {
         viewModelScope.launch {
             communityRes = ApiState.Loading
             communityRes = API.getInstance().getCommunity(form).toApiState()
         }
-    }
-
-    private fun getPosts(
-        form: GetPosts,
-        state: ApiState<GetPostsResponse> = ApiState.Loading,
-    ) {
-        viewModelScope.launch {
-            postsRes = state
-            postsRes = API.getInstance().getPosts(form).toApiState()
-        }
-    }
-
-    fun appendPosts() {
-        viewModelScope.launch {
-            val oldRes = postsRes
-            postsRes =
-                when (oldRes) {
-                    is ApiState.Appending -> return@launch
-                    is ApiState.Holder -> ApiState.Appending(oldRes.data)
-                    else -> return@launch
-                }
-
-            // Update the page cursor before fetching again
-            pageCursor = oldRes.data.next_page
-            nextPage()
-
-            val newRes = API.getInstance().getPosts(getForm()).toApiState()
-
-            postsRes =
-                when (newRes) {
-                    is ApiState.Success -> {
-                        ApiState.Success(
-                            GetPostsResponse(
-                                posts =
-                                    mergePosts(
-                                        oldRes.data.posts,
-                                        newRes.data.posts,
-                                    ),
-                                next_page = newRes.data.next_page,
-                            ),
-                        )
-                    }
-
-                    else -> {
-                        prevPage()
-                        ApiState.AppendingFailure(oldRes.data)
-                    }
-                }
-        }
-    }
-
-    fun resetPosts() {
-        resetPage()
-        getPosts(
-            getForm(),
-        )
-    }
-
-    fun refreshPosts() {
-        resetPage()
-        getPosts(
-            getForm(),
-            ApiState.Refreshing,
-        )
     }
 
     fun followCommunity(
@@ -171,49 +60,6 @@ class CommunityViewModel(communityArg: Either<CommunityId, String>) : ViewModel(
 
                         else -> {}
                     }
-                }
-
-                else -> {}
-            }
-        }
-    }
-
-    fun likePost(form: CreatePostLike) {
-        viewModelScope.launch {
-            likePostRes = ApiState.Loading
-            likePostRes = API.getInstance().createPostLike(form).toApiState()
-
-            when (val likeRes = likePostRes) {
-                is ApiState.Success -> {
-                    updatePost(likeRes.data.post_view)
-                }
-
-                else -> {}
-            }
-        }
-    }
-
-    fun savePost(form: SavePost) {
-        viewModelScope.launch {
-            savePostRes = ApiState.Loading
-            savePostRes = API.getInstance().savePost(form).toApiState()
-            when (val saveRes = savePostRes) {
-                is ApiState.Success -> {
-                    updatePost(saveRes.data.post_view)
-                }
-
-                else -> {}
-            }
-        }
-    }
-
-    fun deletePost(form: DeletePost) {
-        viewModelScope.launch {
-            deletePostRes = ApiState.Loading
-            deletePostRes = API.getInstance().deletePost(form).toApiState()
-            when (val deletePost = deletePostRes) {
-                is ApiState.Success -> {
-                    updatePost(deletePost.data.post_view)
                 }
 
                 else -> {}
@@ -254,37 +100,6 @@ class CommunityViewModel(communityArg: Either<CommunityId, String>) : ViewModel(
         }
     }
 
-    fun updatePost(postView: PostView) {
-        when (val existing = postsRes) {
-            is ApiState.Success -> {
-                val newPosts = findAndUpdatePost(existing.data.posts, postView)
-                val newRes = ApiState.Success(existing.data.copy(posts = newPosts))
-                postsRes = newRes
-            }
-
-            else -> {}
-        }
-    }
-
-    fun markPostAsRead(
-        form: MarkPostAsRead,
-        post: PostView,
-        appState: JerboaAppState,
-    ) {
-        appState.coroutineScope.launch {
-            markPostRes = ApiState.Loading
-            markPostRes = API.getInstance().markPostAsRead(form).toApiState()
-
-            when (markPostRes) {
-                is ApiState.Success -> {
-                    updatePost(post.copy(read = form.read))
-                }
-
-                else -> {}
-            }
-        }
-    }
-
     init {
         communityId = communityArg.fold({ it }, { null })
         communityName = communityArg.fold({ null }, { it })
@@ -303,7 +118,7 @@ class CommunityViewModel(communityArg: Either<CommunityId, String>) : ViewModel(
         )
     }
 
-    private fun getForm(): GetPosts {
+    override fun getForm(): GetPosts {
         return GetPosts(
             community_id = communityId,
             community_name = communityName,
